requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
asgiref==3.5.0
    # via
    #   django
    #   djoser
backports-zoneinfo==0.2.1
    # via django
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.12
    # via requests
coreapi==2.3.3
    # via djoser
coreschema==0.0.4
    # via coreapi
cryptography==36.0.1
    # via social-auth-core
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
django==4.0.2
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cors-headers==3.11.0
    # via -r requirements.in
django-templated-mail==1.1.1
    # via djoser
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==4.8.0
    # via djoser
djoser==2.1.0
    # via -r requirements.in
idna==3.3
    # via requests
itypes==1.2.0
    # via coreapi
jinja2==3.0.3
    # via coreschema
markupsafe==2.1.0
    # via jinja2
oauthlib==3.2.0
    # via
    #   requests-oauthlib
    #   social-auth-core
pycparser==2.21
    # via cffi
pyjwt==2.3.0
    # via
    #   djangorestframework-simplejwt
    #   social-auth-core
python3-openid==3.2.0
    # via social-auth-core
pytz==2021.3
    # via djangorestframework
requests==2.27.1
    # via
    #   coreapi
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.1
    # via social-auth-core
six==1.16.0
    # via social-auth-app-django
social-auth-app-django==4.0.0
    # via djoser
social-auth-core==4.2.0
    # via social-auth-app-django
sqlparse==0.4.2
    # via django
uritemplate==4.1.1
    # via coreapi
urllib3==1.26.8
    # via requests
